zmiany

+++ src/CmsIr/Category/src/Category/Controller/CategoryController.php
@@ -7,6 +7,7 @@ use CmsIr\Category\Model\Category;
 use CmsIr\Dictionary\Form\DictionaryForm;
 use CmsIr\Dictionary\Form\DictionaryFormFilter;
 use CmsIr\Dictionary\Model\Dictionary;
+use CmsIr\File\Model\File;
 use CmsIr\System\Model\MenuTable;
 use CmsIr\System\Util\Inflector;
 use Zend\Mvc\Controller\AbstractActionController;
@@ -17,7 +18,8 @@ use Zend\Json\Json;
 
 class CategoryController extends AbstractActionController
 {
-    protected $uploadDir = 'public/files/category/';
+    protected $uploadDir = 'public/temp_files/category/';
+    protected $destinationUploadDir = 'public/files/category/';
 
     public function listAction()
     {
@@ -64,6 +66,25 @@ class CategoryController extends AbstractActionController
                 $category->exchangeArray($form->getData());
                 $id = $this->getCategoryTable()->save($category);
 
+                $scannedDirectory = array_diff(scandir($this->uploadDir), array('..', '.'));
+                if(!empty($scannedDirectory))
+                {
+                    foreach($scannedDirectory as $file)
+                    {
+                        $mimeType = $this->getFileService()->getMimeContentType($this->uploadDir.'/'.$file);
+
+                        $postFile = new File();
+                        $postFile->setFilename($file);
+                        $postFile->setEntityId($id);
+                        $postFile->setEntityType('Category');
+                        $postFile->setMimeType($mimeType);
+
+                        $this->getFileTable()->save($postFile);
+
+                        rename($this->uploadDir.'/'.$file, $this->destinationUploadDir.'/'.$file);
+                    }
+                }
+
                 $this->getBlockService()->saveBlocks($id, 'Category', $request->getPost()->toArray(), 'title');
 
                 $this->flashMessenger()->addMessage('Kategoria została dodana poprawnie.');
@@ -91,6 +112,7 @@ class CategoryController extends AbstractActionController
         }
 
         $blocks = $this->getBlockService()->getBlocks($category, 'Category');
+        $categoryFiles = $this->getFileTable()->getBy(array('entity_id' => $id, 'entity_type' => 'Category'));
 
         $form = new CategoryForm();
         $form->bind($category);
@@ -105,6 +127,26 @@ class CategoryController extends AbstractActionController
             if ($form->isValid())
             {
                 $this->getCategoryTable()->save($category);
+
+                $scannedDirectory = array_diff(scandir($this->uploadDir), array('..', '.'));
+                if(!empty($scannedDirectory))
+                {
+                    foreach($scannedDirectory as $file)
+                    {
+                        $mimeType = $this->getFileService()->getMimeContentType($this->uploadDir.'/'.$file);
+
+                        $postFile = new File();
+                        $postFile->setFilename($file);
+                        $postFile->setEntityId($id);
+                        $postFile->setEntityType('Category');
+                        $postFile->setMimeType($mimeType);
+
+                        $this->getFileTable()->save($postFile);
+
+                        rename($this->uploadDir.'/'.$file, $this->destinationUploadDir.'/'.$file);
+                    }
+                }
+
                 $this->getBlockService()->saveBlocks($id, 'Category', $request->getPost()->toArray(), 'title');
 
                 $this->flashMessenger()->addMessage('Kategoria została edytowana poprawnie.');
@@ -116,6 +158,7 @@ class CategoryController extends AbstractActionController
         $viewParams = array();
         $viewParams['form'] = $form;
         $viewParams['blocks'] = $blocks;
+        $viewParams['categoryFiles'] = $categoryFiles;
         $viewModel = new ViewModel();
         $viewModel->setVariables($viewParams);
         return $viewModel;
@@ -160,9 +203,10 @@ class CategoryController extends AbstractActionController
         );
     }
 
-    public function uploadAction ()
+    public function uploadFilesMainAction ()
     {
-        if (!empty($_FILES)) {
+        if (!empty($_FILES))
+        {
             $tempFile   = $_FILES['Filedata']['tmp_name'];
             $targetFile = $_FILES['Filedata']['name'];
 
@@ -173,9 +217,36 @@ class CategoryController extends AbstractActionController
             $uniqidFilename = $fileName.'-'.uniqid();
             $targetFile = $uniqidFilename.'.'.$fileExt;
 
-            if(move_uploaded_file($tempFile,$this->uploadDir.$targetFile)) {
+            if(move_uploaded_file($tempFile,$this->destinationUploadDir.$targetFile))
+            {
                 echo $targetFile;
-            } else {
+            } else
+            {
+                echo 0;
+            }
+        }
+        return $this->response;
+    }
+
+    public function uploadFilesAction ()
+    {
+        if (!empty($_FILES))
+        {
+            $tempFile   = $_FILES['Filedata']['tmp_name'];
+            $targetFile = $_FILES['Filedata']['name'];
+            $file = explode('.', $targetFile);
+
+            $fileName = $file[0];
+            $fileExt = $file[1];
+
+            $uniqidFilename = $fileName.'-'.uniqid();
+            $targetFile = $uniqidFilename.'.'.$fileExt;
+
+            if(move_uploaded_file($tempFile,$this->uploadDir.$targetFile))
+            {
+                echo $targetFile;
+            } else
+            {
                 echo 0;
             }
 
@@ -191,7 +262,39 @@ class CategoryController extends AbstractActionController
             $name = $request->getPost('name');
             $filePath = $request->getPost('filePath');
 
-            unlink('./public'.$filePath);
+            if(!empty($id))
+            {
+                $this->getFileTable()->deleteFile($id);
+                unlink('./public'.$filePath);
+
+            } else
+            {
+                unlink('./public'.$filePath);
+            }
+        }
+
+        $jsonObject = Json::encode($params['status'] = 'success', true);
+        echo $jsonObject;
+        return $this->response;
+    }
+
+    public function deletePhotoMainAction()
+    {
+        $request = $this->getRequest();
+        if ($request->isPost()) {
+            $id = $request->getPost('id');
+            $name = $request->getPost('name');
+            $filePath = $request->getPost('filePath');
+
+            if(!empty($id))
+            {
+                $this->getFileTable()->deleteFile($id);
+                unlink('./public'.$filePath);
+
+            } else
+            {
+                unlink('./public'.$filePath);
+            }
         }
 
         $jsonObject = Json::encode($params['status'] = 'success', true);
@@ -213,5 +316,21 @@ class CategoryController extends AbstractActionController
     public function getBlockService()
     {
         return $this->getServiceLocator()->get('CmsIr\System\Service\BlockService');
+    }
+
+    /**
+     * @return \CmsIr\File\Service\FileService
+     */
+    public function getFileService()
+    {
+        return $this->getServiceLocator()->get('CmsIr\File\Service\FileService');
+    }
+
+    /**
+     * @return \CmsIr\File\Model\FileTable
+     */
+    public function getFileTable()
+    {
+        return $this->getServiceLocator()->get('CmsIr\File\Model\FileTable');
     }
 }
